requirements.txt
coverage==4.0.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-dotenv==1.4.1
djangorestframework==3.3.2
djangorestframework-jsonp==1.0.2
enum34==1.1.6
gunicorn==19.6.0
nba_data==1.6
nose==1.3.7
psycopg2==2.6.2
python-coveralls==2.7.0
pytz==2016.10
PyYAML==3.12
requests==2.10.0
sh==1.11
six==1.10.0
static3==0.7.0
whitenoise==2.0.6
